+++ lib/glimmer/dsl/libui/dsl.rb
@@ -40,6 +40,7 @@ module Glimmer
           data_binding
           shine_data_binding
           property
+          string
           control
           shape
         ]